+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Model `fill()` example
+- Model methods:
+  - `isFresh()`
 
 ### Changed